Separate mock tests of internet service and gateway into another files
